:zap: postCellVM 삭제, 함수 이동
